Introduce RangeVectorOperator interface and split RangeVectorSelectorWithTransformation into two types.
